+++ Market/MarketDay/Data/ProgressionModel.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using MarketDay.Utility;
 using StardewModdingAPI;
 
@@ -6,9 +8,11 @@ namespace MarketDay.Data
 {
     public class PrizeLevel
     {
+        public string Name { get; set; }
         public int Gold { get; set; }
         public int Score { get; set; }
         public string Object { get; set; }
+        public string Flavor { get; set; }
         public int Quality { get; set; } = 0;
         public int Stack { get; set; } = 3;
     }
@@ -18,12 +22,60 @@ namespace MarketDay.Data
         public string Name { get; set; }
         public int MarketSize { get; set; }
         public int UnlockAtEarnings { get; set; }
+
+        public int AutoRestock { get; set; } = 4;
+
+        public int ShopSize { get; set; } = 9;
+
+        public double PriceMultiplier { get; set; } = 1;
+        
         public List<PrizeLevel> Prizes { get; set; }
+
+        public PrizeLevel PrizeForEarnings(int gold)
+        {
+            var eligiblePrizes = Prizes.Where(p => p.Score == 0 && p.Gold <= gold).OrderBy(p => p.Gold);
+            return eligiblePrizes.Any() ? eligiblePrizes.Last() : null;
+        }
+        
     }
     
     public class ProgressionModel
     {
         public List<ProgressionLevel> Levels { get; set; }
+
+        internal ProgressionLevel CurrentLevel
+        {
+            get
+            {
+                ProgressionLevel highestUnlocked = null;
+                var gold = MarketDay.GetSharedValue(MarketDay.TotalGoldKey);
+                foreach (var level in Levels.Where(level => level.UnlockAtEarnings <= gold)) highestUnlocked = level;
+                return highestUnlocked;
+            }
+        }
+
+        internal int AutoRestock =>
+            Math.Max(0, 
+                MarketDay.Config.Progression 
+                ? CurrentLevel.AutoRestock
+                : MarketDay.Config.RestockItemsPerHour
+                );
+
+        internal int ShopSize =>
+            Math.Max(1, Math.Min(9, 
+                MarketDay.Config.Progression 
+                ? CurrentLevel.ShopSize
+                : 9
+                ));
+
+        internal double PriceMultiplier =>
+            Math.Max(1, Math.Min(4, 
+                MarketDay.Config.Progression 
+                ? CurrentLevel.PriceMultiplier
+                : 1
+            ));
+
+        internal int GoldTarget => CurrentLevel?.Prizes.Where(p => p.Score==0).OrderBy(p => p.Gold).First().Gold ?? 0;
 
         internal void CheckItems()
         {