+++ Market/MarketDay/ModConfig.cs
@@ -5,6 +5,10 @@ namespace MarketDay
 {
     internal class ModConfig
     {
+        public bool Progression { get; set; } = true;
+        
+        public bool OnNextDayIfFestival { get; set; } = true;
+
         public int DayOfWeek { get; set; } = 6;
         public int OpeningTime { get; set; } = 8;
         public int ClosingTime { get; set; } = 18;
@@ -17,11 +21,12 @@ namespace MarketDay
         public bool ReceiveMessages { get; set; } = true;
         public bool PeekIntoChests { get; set; }
         public bool RuinTheFurniture { get; set; }
-        public Dictionary<string, bool> ShopsEnabled = new Dictionary<string, bool>();
-        public bool VerboseLogging { get; set; } = false;
+        public Dictionary<string, bool> ShopsEnabled = new();
+        public bool VerboseLogging { get; set; }
         public bool NPCVisitors { get; set; } = true;
         public bool NPCRescheduling { get; set; } = true;
-        public bool DebugKeybinds { get; set; } = false;
+        public bool AlwaysMarketDay { get; set; }
+        public bool DebugKeybinds { get; set; }
         public SButton OpenConfigKeybind { get; set; } = SButton.V;
         public SButton WarpKeybind { get; set; } = SButton.Z;
         public SButton ReloadKeybind { get; set; } = SButton.R;